even TCP answers can only be 65535 bytes long! Plus, we misinterpreted the EDNS0 maximum reply length limit. Spotted by Thor Spruyt.
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@2150 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/dnswriter.cc
@@ -278,7 +278,7 @@ void DNSPacketWriter::getRecords(string& records)
   records.assign(d_content.begin() + d_sor, d_content.end());
 }
 
-uint16_t DNSPacketWriter::size()
+uint32_t DNSPacketWriter::size()
 {
   return d_content.size() + d_stuff + d_record.size();
 }
